Replaced Models with Cache
Consoldiated database and cache modules and eliminated Models package
