+++ delirium/dns/__init__.py
@@ -1,2 +1,3 @@
 from .fakednsserver import FakeDNSServer
 from .fakeresolver import FakeResolver
+from .cache import *
